degenerate test for retry loop, factoring the forAllFuture utility
