+++ triggers/tests/BUILD.bazel
@@ -122,6 +122,7 @@ da_scala_library(
                 "FuncTestsWallClock",
                 "Jwt",
                 "Tls",
+                "RunnerSpec",
                 "UnfoldStateSpec",
             ]
         ],
@@ -155,6 +156,7 @@ da_scala_library(
             "//ledger/test-common",
             "//libs-scala/ports",
             "//libs-scala/resources",
+            "//libs-scala/scalatest-utils",
             "//triggers/runner:trigger-runner-lib",
             "@maven//:com_typesafe_akka_akka_stream_2_12",
             "@maven//:org_scalacheck_scalacheck_2_12",
