+++ triggers/tests/src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/UnfoldStateSpec.scala
@@ -7,24 +7,21 @@ package test
 import akka.stream.ClosedShape
 import akka.stream.scaladsl.{GraphDSL, Keep, RunnableGraph, Sink, Source}
 import com.daml.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import org.scalacheck.Gen.listOfN
-import org.scalacheck.Arbitrary
+import com.daml.scalatest.AsyncForAll
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
-import org.scalatest.Assertion
 import org.scalatest.{AsyncWordSpec, Matchers}
-import scalaz.{\/, -\/, \/-, Applicative, Monoid}
+import scalaz.{\/, -\/, \/-}
 import scalaz.std.list._
 import scalaz.std.scalaFuture._
 import scalaz.syntax.apply.^
 import scalaz.syntax.bifunctor._
 import scalaz.syntax.traverse._
 
-import scala.concurrent.Future
-
 class UnfoldStateSpec
     extends AsyncWordSpec
     with Matchers
     with GeneratorDrivenPropertyChecks
+    with AsyncForAll
     with AkkaBeforeAndAfterAll {
   import UnfoldState._
 
@@ -45,7 +42,7 @@ class UnfoldStateSpec
   }
 
   "flatMapConcat" should {
-    "do as built-in flatMapConcat would" in forAllFuture(trials = 10) { run: List[List[Int]] =>
+    "do as built-in flatMapConcat would" in forAllAsync(trials = 10) { run: List[List[Int]] =>
       val flattened = run.flatten
       var escape = (0, 0)
       Source(run)
@@ -66,25 +63,24 @@ class UnfoldStateSpec
   }
 
   "flatMapConcatStates" should {
-    "emit every state after the list elements" in forAllFuture(trials = 10) {
-      run: List[List[Int]] =>
-        val (_, expected) = run.mapAccumL(0) { (sum, ns) =>
-          val newSum = sum + ns.sum
-          (newSum, (ns map \/.right) :+ -\/(newSum))
+    "emit every state after the list elements" in forAllAsync(trials = 10) { run: List[List[Int]] =>
+      val (_, expected) = run.mapAccumL(0) { (sum, ns) =>
+        val newSum = sum + ns.sum
+        (newSum, (ns map \/.right) :+ -\/(newSum))
+      }
+      Source(run)
+        .via(flatMapConcatStates(0) { (sum, ns) =>
+          fromLinearSeq(ns) leftMap (_ => sum + ns.sum)
+        })
+        .runWith(Sink.seq)
+        .map { ran =>
+          ran should ===(expected.flatten)
         }
-        Source(run)
-          .via(flatMapConcatStates(0) { (sum, ns) =>
-            fromLinearSeq(ns) leftMap (_ => sum + ns.sum)
-          })
-          .runWith(Sink.seq)
-          .map { ran =>
-            ran should ===(expected.flatten)
-          }
     }
   }
 
   "flatMapConcatNode" should {
-    "emit same elements as flatMapConcat" in forAllFuture(trials = 10) { run: List[List[Int]] =>
+    "emit same elements as flatMapConcat" in forAllAsync(trials = 10) { run: List[List[Int]] =>
       val (_, concatPairs) = run.mapAccumL(0) { (sum, ns) =>
         val newSum = sum + ns.sum
         (newSum, (ns, newSum))
@@ -107,15 +103,5 @@ class UnfoldStateSpec
         ns should ===(concatPairs flatMap (_._1))
       }
     }
-  }
-
-  private[this] def forAllFuture[A](trials: Int)(f: A => Future[Assertion])(
-      implicit A: Arbitrary[A]): Future[Assertion] = {
-    val runs = listOfN(trials, A.arbitrary).sample
-      .getOrElse(sys error "random Gen failed")
-
-    implicit val assertionMonoid: Monoid[Future[Assertion]] =
-      Monoid liftMonoid (Applicative[Future], Monoid instance ((_, result) => result, succeed))
-    runs foldMap f
   }
 }
